Profile Navigation and refactored
