+++ Threads/Core/Profile/Components/ProfileHeaderView.swift
@@ -1,0 +1,50 @@
+//
+//  ProfileHeaderView.swift
+//  Threads
+//
+//  Created by aykut ipek on 18.09.2023.
+//
+
+import SwiftUI
+
+struct ProfileHeaderView: View {
+    var user: User?
+    
+    init(user: User?) {
+        self.user = user
+    }
+    var body: some View {
+        HStack(alignment: .top) {
+            VStack(alignment: .leading, spacing: 12) {
+                
+                VStack(alignment: .leading, spacing: 4) {
+                    Text(user?.fullname ?? "")
+                        .font(.title2)
+                        .fontWeight(.semibold)
+                    
+                    Text(user?.username ?? "")
+                        .font(.subheadline)
+                }
+
+                if let bio = user?.bio {
+                    Text(bio)
+                        .font(.footnote)
+                }
+
+                Text("2 Followers")
+                    .font(.caption)
+                    .foregroundColor(.gray)
+            }
+            
+            Spacer()
+            
+            CircularProfileImageView()
+        }
+    }
+}
+
+struct ProfileHeaderView_Previews: PreviewProvider {
+    static var previews: some View {
+        ProfileHeaderView(user: dev.user)
+    }
+}
