+++ Threads/Core/Profile/Components/UserContentListView.swift
@@ -1,0 +1,59 @@
+//
+//  UserContentListView.swift
+//  Threads
+//
+//  Created by aykut ipek on 18.09.2023.
+//
+
+import SwiftUI
+
+struct UserContentListView: View {
+    @State private var selectedFilter: ProfileThreadFilter = .threads
+    @Namespace var animation
+    private var filterBarWidth: CGFloat {
+        let count = CGFloat(ProfileThreadFilter.allCases.count)
+        return UIScreen.main.bounds.width / count - 16
+    }
+    
+    var body: some View {
+        VStack {
+            HStack {
+                ForEach(ProfileThreadFilter.allCases) { filter in
+                    VStack {
+                        Text(filter.title)
+                            .font(.subheadline)
+                            .fontWeight(selectedFilter == filter ? .semibold : .regular)
+                        
+                        if selectedFilter == filter {
+                            Rectangle()
+                                .foregroundColor(.primary)
+                                .frame(width: filterBarWidth, height: 1)
+                                .matchedGeometryEffect(id: "item", in: animation)
+                        } else {
+                            Rectangle()
+                                .foregroundColor(.clear)
+                                .frame(width: filterBarWidth, height: 1)
+                        }
+                    }
+                    .onTapGesture {
+                        withAnimation(.spring()) {
+                            selectedFilter = filter
+                        }
+                    }
+                }
+            }
+            LazyVStack {
+                ForEach(0 ... 10, id: \.self) { thread in
+                    ThreadCell()
+                }
+            }
+        }
+        .padding(.vertical, 8)
+    }
+}
+
+struct UserContentListView_Previews: PreviewProvider {
+    static var previews: some View {
+        UserContentListView()
+    }
+}
